Adding city weather.

+++ LearningUnit2/GeoWeather/app/src/main/java/za/ac/iie/opsc/geoweather/model/location/TimeZone.kt
@@ -7,7 +7,7 @@ data class TimeZone (
 
   @SerializedName("Code"             ) var Code             : String?  = null,
   @SerializedName("Name"             ) var Name             : String?  = null,
-  @SerializedName("GmtOffset"        ) var GmtOffset        : Int?     = null,
+  @SerializedName("GmtOffset"        ) var GmtOffset        : Float?   = null,
   @SerializedName("IsDaylightSaving" ) var IsDaylightSaving : Boolean? = null,
   @SerializedName("NextOffsetChange" ) var NextOffsetChange : String?  = null
